Updates gulpfiles to compile ts->js->rct
