+++ .gitignore
@@ -1,4 +1,5 @@
 node_modules
 lib
+temp
 src/MessageInput
 .DS_Store
