+++ .npmignore
@@ -1,3 +1,4 @@
 src
+temp
 demo
 *.DS_Store
